Tests de los enemigos

+++ src/server_src/personaje/personaje.h
@@ -23,7 +23,7 @@ protected:
     uint32_t id_jugador;
     uint32_t posicion_x;
     uint32_t posicion_y;
-    int vida; //DEBE SER INT PARA PODER SER NEGATIVO
+    int8_t vida; //DEBE SER INT PARA PODER SER NEGATIVO
     bool esta_quieto;
     uint32_t ancho;
     uint32_t alto;
